Added Item detection and Inventory
Added an item item detection system using a simple hit scan from where the player is looking which will destroy the object and add the item to the players inventory. I then also created a player class which can manage any player stats. Finally there is now an item class to add things to the inventory. Currently only contains name and description but can be updated to use other parameters such as tags for ammo, weapons, and key items.