Fixed bug when adding items
Items weren't being added to the inventory properly with items not stacking and instead every instance of a collected item would be detected as the a unique item. This has been changed so that now if there is an existing item called "cube" and another "cube" is added there will be an item called "cube" with a count of 2 as opposed to two items with a count of 1.